+++ CHANGELOG.md
@@ -106,6 +106,7 @@
 * Refactored `BoundingBox.extract_from_image()`. Improved some code fragments that looked wrong.
 * Changed `KeypointsOnImage.draw_on_image()` to raise an error if an input image has shape `(H,W)`.
 * Changed `KeypointsOnImage.draw_on_image()` to handle single-number inputs for `color`.
+* Added function `imgaug.imgaug.flatten()` that flattens nested lists/tuples.
 
 
 # 0.2.8